+++ Paste-relevant.ps1
@@ -14,10 +14,15 @@ $srcPaths = @($SAP, $MOC, $Downloads)
 
 #Iterating through the array
 foreach ($obj in $srcPaths) {
-    # checks if the path, like "Path/To/SAP" exists and if not it will created as a dir
-    if(-not(Test-Path "$obj[1]")) {
-        New-Item -Path "$obj[1]" -ItemType Directory
+    # checks if the folder is in OneDrive (and software is installed)
+    if(Test-Path "$OneDrive\$obj[0]" -And Test-Path "$obj[1]") {
+        # checks if the path, like "Path/To/SAP" exists and if not it will created as a dir
+        # side note: if checked previous this if-condition can be commented out
+        if(-not(Test-Path "$obj[1]")) {
+            New-Item -Path "$obj[1]" -ItemType Directory
+        }
+        # copies everything the users OneDrive back to said app user-config folders
+        Copy-Item -Path "$OneDrive\$obj[0]" -Destination "$obj[1]" -Recurse
+
     }
-    # copies everything the users OneDrive back to said app user-config folders
-    Copy-Item -Path "$OneDrive\$obj[0]" -Destination "$obj[1]" -Recurse
 }
